requirements.txt
# Django and REST framework
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication
django-allauth==0.60.1
djangorestframework-simplejwt==5.3.1

# Web3 and Blockchain
web3==6.15.1
eth-account==0.10.0
eth-utils==2.3.1

# Quantum-resistant cryptography
liboqs==0.9.0
#pyoqs==0.9.0

# AI and ML
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
matplotlib==3.8.2
joblib==1.3.2

# Utilities
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0

